+++ lib/pal.dart
@@ -4,3 +4,4 @@ export 'package:pal/src/ui/client/helper_orchestrator.dart';
 export 'package:pal/src/pal.dart' show Pal;
 export 'package:pal/src/pal_navigator_observer.dart' show PalNavigatorObserver;
 export 'package:pal/src/in_app_user_manager.dart' show InAppUserManager;
+export 'package:pal/src/services/app_events.dart' show PalEvents;